Modifies header
Modifies header to implement new bootstrap style. NOTE: The navbar section links are not intended to be permanent but are left up there so we still have easy access to the other pages. Once we decide what to do with the PhysioNet and Critical Care pages, then we can modify those respectively.