Updates People Page
Updates the People page to the new bootstrap template. Moves the links to each person to a sidebar. Also adds new links to specific research category. Adds section for graduate students that I think Felipe maybe forgot to add since it is in the render template function but not actually on the template so I added it. Also adds pictures for the collaborators to maintain the same form. I will open an issue about the missing photos since Roger does not like that.