Changes footer style
Moves logos from the bottom of the home page to the bottom of the footer so it can be seen from every page. Changes footer background to white to follow template style more consistently. Adds a Read More option to the publications list to direct the user to the publications page to view more content if desired. Adds links to each logo to direct the user to their respective site.