Adds news feature
Adds a new tab in the navbar for news items. I chose to stick with the same style as the publications page since it is a similar feature. The current news items should probably be changed but I just wanted a wide variety of dates to make sure the date pagination works. The sample news items are taken from the PhysioNet news fixtures. Currently, I assign unique post numbers to each news item to refer to them later but feel free to suggest other options if you think there is a better way out there.

Removes pdb