Adds news items to front page
Adds news items to front page below the lab welcome image. The user can adjust which news items are selected for the front page using the associated news.yml file. There is no limit to the number of news items which can be displayed. With each news item, the user can create a header which is a quick summary of the news item used on the news display to grab the attention of the reader. Content is truncated at 150 characters to avoid content which is too long. No links were added though may be added in a future pull request.