Fixes sidebar squash #60
Fixes the sidebar being squashed when the window width is reduced. Previously, the sidebar would compress to a very small width which would not go away even if it wrapped to a new line. Now, the sidebar keeps a minimum width of 150px while expanding if wrapped to a new line. Fixes part of #60.