Adds news scrolling #60
Adds scrolling functionality to the front page news items. Links to the news page are removed since they did not lead to the correct page. News box items are set to a rigid height and the inside text is set to scroll if needed while keeping content such as the title and date static. Fixes part of #60.